Darkmoon faire draft - part 2.1
Structural change
