+++ src/server/scripts/DarkmoonIsland/DarkmoonIsland.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 ShadowCore
+ * This file is part of the TrinityCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -252,162 +252,6 @@ public:
             player->CastSpell(player, 103740, false);
         }
 
-        return true;
-    }
-};
-
-enum JESSICA_GOSSIP
-{
-    GOSSIP_BUTTON_1 = 13012,
-    GOSSIP_BUTTON_1_ID = 0,
-    GOSSIP_BUTTON_2 = 13012,
-    GOSSIP_BUTTON_2_ID = 1,
-    GOSSIP_BUTTON_3 = 13013,
-    GOSSIP_BUTTON_3_ID = 0,
-    
-};
-
-// 54485 - Jessica Rogers
-class npc_jessica_rogers : public CreatureScript
-{
-public:
-    npc_jessica_rogers() : CreatureScript("npc_jessica_rogers") { }
-
-    struct npc_jessica_rogersAI : public ScriptedAI
-    {
-        npc_jessica_rogersAI(Creature* creature) : ScriptedAI(creature) { }
-
-        bool OnGossipHello(Player* player) override
-        {
-            if (me->IsQuestGiver())
-                player->PrepareQuestMenu(me->GetGUID());
-
-            AddGossipItemFor(player, GOSSIP_BUTTON_1, GOSSIP_BUTTON_1_ID, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 1);
-            SendGossipMenuFor(player, 18284, me->GetGUID());
-
-            if (player->GetQuestStatus(QUEST_TARGET_TURTLE) == QUEST_STATUS_INCOMPLETE)
-            {
-                AddGossipItemFor(player, GOSSIP_BUTTON_2, GOSSIP_BUTTON_2_ID, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 2);
-                SendGossipMenuFor(player, 53027, me->GetGUID());
-            }
-
-            player->PlayerTalkClass->SendGossipMenu(player->GetGossipTextId(me), me->GetGUID());
-            return true;
-        }
-
-        bool OnGossipSelect(Player* player, uint32 /*menuId*/, uint32 gossipListId) override
-        {
-            uint32 const action = player->PlayerTalkClass->GetGossipOptionAction(gossipListId);
-            ClearGossipMenuFor(player);
-
-            switch (action)
-            {
-            case GOSSIP_ACTION_INFO_DEF + 1:
-                AddGossipItemFor(player, GOSSIP_BUTTON_3, GOSSIP_BUTTON_3_ID, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 3);
-                break;
-            case GOSSIP_ACTION_INFO_DEF + 2:
-                if (player->HasItemCount(ITEM_DARKMOON_TOKEN))
-                {
-                    CloseGossipMenuFor(player);
-
-                    player->DestroyItemCount(ITEM_DARKMOON_TOKEN, 1, true);
-                    player->RemoveAurasByType(SPELL_AURA_MOUNTED);
-
-                    player->AddAura(SPELL_RINGTOSS_ENABLE, player);
-                    player->SetPower(POWER_ALTERNATE_POWER, 10);
-
-                    return true;
-                }
-                break;
-            case GOSSIP_ACTION_INFO_DEF + 3:
-                AddGossipItemFor(player, GOSSIP_BUTTON_1, GOSSIP_BUTTON_1_ID, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 1);
-                AddGossipItemFor(player, GOSSIP_BUTTON_2, GOSSIP_BUTTON_2_ID, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 2);
-                break;
-            }
-            return true;
-        }
-    };
-
-    CreatureAI* GetAI(Creature* creature) const override
-    {
-        return new npc_jessica_rogersAI(creature);
-    }
-};
-
-//
-class spell_ring_toss : public SpellScriptLoader
-{
-public:
-    spell_ring_toss() : SpellScriptLoader("spell_ring_toss") {}
-
-    class spell_ring_toss_SpellScript : public SpellScript
-    {
-        PrepareSpellScript(spell_ring_toss_SpellScript);
-
-        SpellCastResult CheckRequirement()
-        {
-            if (GetCaster()->GetTypeId() != TYPEID_PLAYER)
-                return SPELL_FAILED_CANT_DO_THAT_RIGHT_NOW;
-
-            return SPELL_CAST_OK;
-        }
-
-        void HandleDummy(SpellEffIndex /*effIndex*/)
-        {
-            Unit* caster = GetCaster();
-            Player* player = GetCaster()->ToPlayer();
-
-            WorldLocation* loc = GetHitDest();
-
-            Creature* dubenko = caster->FindNearestCreature(54490, 100.0f, true);
-
-            //dubenko->CastSpell(dubenko, 101737, false);
-            if (dubenko && dubenko->IsWithinDist2d(loc, 1.2f))
-            {
-                if (dubenko->HasAura(SPELL_RINGTOSS_TURTLE_CIRCLE_1))
-                {
-                    if (dubenko->HasAura(SPELL_RINGTOSS_TURTLE_CIRCLE_2))
-                        dubenko->CastSpell(dubenko, SPELL_RINGTOSS_TURTLE_CIRCLE_3, true);
-                    else
-                        dubenko->CastSpell(dubenko, SPELL_RINGTOSS_TURTLE_CIRCLE_2, true);
-                }
-                else
-                    dubenko->CastSpell(dubenko, SPELL_RINGTOSS_TURTLE_CIRCLE_1, true);
-
-                caster->CastSpell(dubenko, SPELL_RINGTOSS_HIT, false);
-            }
-
-            if (player)
-            {
-                if (player->GetReqKillOrCastCurrentCount(QUEST_TARGET_TURTLE, 54495) >= 3)
-                    player->RemoveAurasDueToSpell(SPELL_RINGTOSS_ENABLE);
-            }            
-        }
-                        
-
-        void Register()
-        {
-            OnEffectHitTarget += SpellEffectFn(spell_ring_toss_SpellScript::HandleDummy, EFFECT_0, SPELL_EFFECT_DUMMY);
-            OnCheckCast += SpellCheckCastFn(spell_ring_toss_SpellScript::CheckRequirement);
-        }
-    };
-
-    SpellScript* GetSpellScript() const
-    {
-        return new spell_ring_toss_SpellScript();
-    }
-};
-
-// at_target_turtle 7338
-class at_target_turtle : public AreaTriggerScript
-{
-public:
-    at_target_turtle() : AreaTriggerScript("at_target_turtle") { }
-
-    bool OnTrigger(Player* player, AreaTriggerEntry const* /*areaTrigger*/)
-    {
-        if (!player->IsGameMaster())
-            player->CastSpell(player, 109972, true);
         return true;
     }
 };
@@ -696,227 +540,6 @@ public:
     }
 };
 
-enum HeShootsSpells
-{
-    SPELL_HESHOOTS_CRACKSHOT_ENABLE = 101871,
-    SPELL_HESHOOTS_SHOOT = 101872,
-    SPELL_HESHOOTS_TARGET_INDICATOR = 101010,
-    SPELL_HESHOOTS_TARGET_INDICATOR_VISUAL = 43313,
-    SPELL_HESHOOTS_KILL_CREDIT = 43300,
-    SPELL_HESHOOTS_KILL_CREDIT_BONUS = 101012,
-};
-
-class npc_rinling_darkmoon : public CreatureScript
-{
-public:
-    npc_rinling_darkmoon() : CreatureScript("npc_rinling_darkmoon") { }
-
-    struct npc_rinling_darkmoonAI : public ScriptedAI
-    {
-        npc_rinling_darkmoonAI(Creature* creature) : ScriptedAI(creature)
-        {
-            Reset();
-        }
-
-        uint32 nextQuoteTimer;
-        uint8 nextQuote;
-
-        uint32 nextTargetChangeTimer;
-        int8 targetIndex;
-        ObjectGuid targetList[3];
-
-        void Reset()
-        {
-            nextQuoteTimer = urand(1, 60) * 1000;
-            nextQuote = 0;
-            nextTargetChangeTimer = 0;
-            targetIndex = -1;
-        }
-
-        void UpdateAI(const uint32 diff)
-        {
-            if (nextTargetChangeTimer <= diff)
-            {
-                if (targetIndex < 0 || targetIndex >= 3)
-                {
-                    std::list<Creature*> crList;
-                    GetCreatureListWithEntryInGrid(crList, me, 24171, 100.0f);
-                    int i = 0;
-                    for (std::list<Creature*>::const_iterator itr = crList.begin(); itr != crList.end(); ++itr)
-                    {
-                        targetList[i++] = (*itr)->GetGUID();
-                        if (i >= 3)
-                            break;
-                    }
-                }
-
-                int8 chosen = -1;
-                do
-                {
-                    // randomly choose one of targets
-                    chosen = urand(0, 2);
-                } while (chosen == targetIndex && urand(0, 1) == 0); // the same target could be chosen with only 50% chance (+-)
-
-                targetIndex = chosen;
-
-                Creature* creature_bunny = ObjectAccessor::GetCreature(*me, targetList[chosen]);
-                if (creature_bunny)
-                    creature_bunny->CastSpell(creature_bunny, SPELL_HESHOOTS_TARGET_INDICATOR, true);
-
-                nextTargetChangeTimer = 6500 + urand(0, 1000);
-            }
-            else
-                nextTargetChangeTimer -= diff;
-        }
-
-        bool OnGossipHello(Player* player) override
-        {
-            if (me->IsQuestGiver())
-                player->PrepareQuestMenu(me->GetGUID());
-
-            AddGossipItemFor(player, 6225, 0, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 1);
-
-            if (player->GetQuestStatus(QUEST_HE_SHOOTS_HE_SCORES) == QUEST_STATUS_INCOMPLETE)
-                AddGossipItemFor(player, 6225, 2, GOSSIP_SENDER_MAIN, GOSSIP_ACTION_INFO_DEF + 2);
-
-            player->PlayerTalkClass->SendGossipMenu(player->GetGossipTextId(me), me->GetGUID());
-            return true;
-        }
-
-        bool OnGossipSelect(Player* player, uint32 /*uiSender*/, uint32 gossipListId) override
-        {
-            uint32 const action = player->PlayerTalkClass->GetGossipOptionAction(gossipListId);
-            ClearGossipMenuFor(player);
-
-            switch (action)
-            {
-                // Info
-            case GOSSIP_ACTION_INFO_DEF + 1:
-                    player->PlayerTalkClass->ClearMenus();
-                    SendGossipMenuFor(player, 18353, me->GetGUID());
-                    return OnGossipHello(player);
-                break;
-                // Ready to play
-            case  GOSSIP_ACTION_INFO_DEF + 2:
-                if (player->HasItemCount(ITEM_DARKMOON_TOKEN))
-                {
-                    CloseGossipMenuFor(player);
-
-                    player->DestroyItemCount(ITEM_DARKMOON_TOKEN, 1, true);
-                    player->CastSpell(player, SPELL_HESHOOTS_CRACKSHOT_ENABLE, true);
-
-                    int16 progress = player->GetReqKillOrCastCurrentCount(QUEST_HE_SHOOTS_HE_SCORES, 54231);
-                    if (progress > 0)
-                        player->SetPower(POWER_ALTERNATE_POWER, progress);
-
-                    return true;
-                }
-                break;
-            }
-
-            return false;
-        }
-    };
-
-    CreatureAI* GetAI(Creature* creature) const
-    {
-        return new npc_rinling_darkmoonAI(creature);
-    }
-};
-
-// spell - 101872
-class spell_heshoots_shoot_hit : public SpellScriptLoader
-{
-public:
-    spell_heshoots_shoot_hit() : SpellScriptLoader("spell_heshoots_shoot_hit") { }
-
-    class spell_heshoots_shoot_hit_SpellScript : public SpellScript
-    {
-        PrepareSpellScript(spell_heshoots_shoot_hit_SpellScript);
-
-        void HandleDummy(SpellEffIndex /*effIndex*/)
-        {
-            if (!GetCaster())
-                return;
-
-            Unit* target = GetHitUnit();
-
-            if (!target)
-                return;
-
-            if (Aura* shootAura = target->GetAura(SPELL_HESHOOTS_TARGET_INDICATOR))
-            {
-                target->CastSpell(GetCaster(), SPELL_HESHOOTS_KILL_CREDIT, true);
-
-                Player* pl = GetCaster()->ToPlayer();
-
-                if (shootAura->GetMaxDuration() - shootAura->GetDuration() < 1000)
-                {
-                    target->CastSpell(GetCaster(), SPELL_HESHOOTS_KILL_CREDIT_BONUS, true);
-                    AchievementEntry const* achiev = sAchievementStore.LookupEntry(6022);
-                    if (pl)
-                        pl->CompletedAchievement(achiev);
-                }
-
-                if (pl)
-                {
-                    if (pl->GetReqKillOrCastCurrentCount(QUEST_HE_SHOOTS_HE_SCORES, 54231) >= 25)
-                        pl->RemoveAurasDueToSpell(SPELL_HESHOOTS_CRACKSHOT_ENABLE);
-                }
-            }
-        }
-
-        void Register()
-        {
-            OnEffectHitTarget += SpellEffectFn(spell_heshoots_shoot_hit_SpellScript::HandleDummy, EFFECT_0, SPELL_EFFECT_DUMMY);
-        }
-    };
-
-    SpellScript* GetSpellScript() const
-    {
-        return new spell_heshoots_shoot_hit_SpellScript();
-    }
-};
-
-// spell - 101010
-class spell_heshoots_indicator : public SpellScriptLoader
-{
-public:
-    spell_heshoots_indicator() : SpellScriptLoader("spell_heshoots_indicator") { }
-
-    class spell_heshoots_indicator_AuraScript : public AuraScript
-    {
-        PrepareAuraScript(spell_heshoots_indicator_AuraScript);
-
-        void EffectApply(AuraEffect const* /*aurEff*/, AuraEffectHandleModes /*modes*/)
-        {
-            if (!GetUnitOwner())
-                return;
-
-            GetUnitOwner()->CastSpell(GetUnitOwner(), SPELL_HESHOOTS_TARGET_INDICATOR_VISUAL, true);
-        }
-
-        void EffectRemove(AuraEffect const* /*aurEff*/, AuraEffectHandleModes /*modes*/)
-        {
-            if (!GetUnitOwner())
-                return;
-
-            GetUnitOwner()->RemoveAurasDueToSpell(SPELL_HESHOOTS_TARGET_INDICATOR_VISUAL);
-        }
-
-        void Register()
-        {
-            OnEffectApply += AuraEffectApplyFn(spell_heshoots_indicator_AuraScript::EffectApply, EFFECT_0, SPELL_AURA_DUMMY, AURA_EFFECT_HANDLE_REAL);
-            OnEffectRemove += AuraEffectApplyFn(spell_heshoots_indicator_AuraScript::EffectRemove, EFFECT_0, SPELL_AURA_DUMMY, AURA_EFFECT_HANDLE_REAL);
-        }
-    };
-
-    AuraScript* GetAuraScript() const
-    {
-        return new spell_heshoots_indicator_AuraScript();
-    }
-};
-
 void AddSC_darkmoon_island()
 {
     new npc_fire_juggler_darkmoon();
@@ -924,12 +547,7 @@ void AddSC_darkmoon_island()
     new spell_darkmoon_carousel_whee();
     new spell_darkmoon_staging_area_teleport();
 
-    new item_darkmoon_faire_fireworks();
-
-    //ring toss
-    new npc_jessica_rogers();
-    new spell_ring_toss();
-    new at_target_turtle();
+    new item_darkmoon_faire_fireworks();   
 
     new spell_gen_repair_damaged_tonk();
     new spell_gen_shoe_baby();
@@ -937,9 +555,4 @@ void AddSC_darkmoon_island()
     new spell_heal_injuried_carnie();
     new spell_put_up_darkmoon_banner();
     new spell_darkmoon_deathmatch();
-
-    //shoot gallery
-    new npc_rinling_darkmoon();
-    new spell_heshoots_shoot_hit();
-    new spell_heshoots_indicator();
 };
