+++ src/server/scripts/DarkmoonIsland/DarkmoonIsland.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 ShadowCore
+ * This file is part of the TrinityCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
